requirements.txt
fire>=0.5.0
PyGithub>=2.1.1
